Update README.md
This commit introduces the first version of the WinUtilsAutomator project. It includes a set of batch scripts designed to automate the installation of Windows in both UEFI and BIOS modes. The scripts handle disk partitioning, applying Windows images with the DISM tool, and configuring bootloaders. This project aims to simplify the Windows installation process for users and will be expanded with additional utilities for system configuration in future releases.