Update README.md
This commit updates the README.md file to introduce WinBatchInstaller as a sub-project within the WinUtilsAutomator repository. The file now includes detailed information about the batch scripts included in WinBatchInstaller, such as win-installer.bat, disk-partitioning.bat, and dism.bat, which automate the Windows installation process in both UEFI and BIOS modes. The README also includes installation instructions, usage guidelines, and contributing details for both the main project and the sub-project. A link to the DOCUMENTATION.md file is provided for users seeking more in-depth setup and usage instructions.